+++ app/src/features/rules/screens/rulesList/components/RulesList/components/RulesTable/constants.ts
