+++ app/src/features/rules/screens/rulesList/components/RulesList/components/RulesTable/utils/index.ts
@@ -1,2 +1,3 @@
 export { convertToArray } from "./rules";
 export { recordsToContentTableRecordsAdapter as recordsToContentTableDataAdapter } from "./rules";
+export { importSampleRules } from "./rules";
